smurf: More mapping simplifications to avoid focal plane distyortion issues
Transforming positions using back-to-back PolyMaps (as per focal plane
distortion) can cause numerical errors due to the inverse PolyMap not
being exactly correct. These errors get big as you go outside the domain
of the subarray to which the polymap applies. SO it is betyter to cancel
out back-to-back PolyMap pairs wherever possible using Mapping simplification
so that their transformations are not used.