smurf: Use matrix inversion to calculate stare Q,U and I values
The previous correlation method of calculating Q,U and I assumed that
trig functions of the analyser angle summed to zero over the course of
one block of time slices. This was only very approximately true, even
with attempt to ensure that each block started and ended at the same
angle.

The new method uses the fact that we know what the analyser angles are
and so can estimate what these non-zero sums actually are (rather than
just assuming they are zero). The new method is implemented as a matrix
inversion, which is mathematically very close to a least squares fit.

   Ianal = 0.5* (I + Q*cos(2.phi) + U*sin(2.phi) )

   2*sum( Ianal*cos(2.phi) ) = I*sum(cos(2.phi)) + Q*sum(cos(2.phi)*cos(2.phi)) + U*sum(sin(2.phi)*cos(2.phi))
   2*sum( Ianal*sin(2.phi) ) = I*sum(sin(2.phi)) + Q*sum(cos(2.phi)*sin(2.phi)) + U*sum(sin(2.phi)*sin(2.phi))
   2*sum( Ianal )            = I*N               + Q*sum(cos(2.phi))            + U*sum(sin(2.phi))

The above equations are now formed and inverted to give values for I, Q and U.