create and install HTML summary of Starlink components.
This uses the existing 'make componentset.html' to create a newly
updated html summary of the libraries, applications and docs built by
Starlink, based on the componentset.xml file. The make world command
will now create this file, and install it into the docs tree as
starlinksummary.html (along with its css).