smurf: Remove restriction that IP ref image must is aligned with output map
pol2scan does not produce a total intensity image (only Q and U images),
and so the IP ref map must be created from some other observation, which
will normally be a standard daisy. But this makes it hard to ensure that
the IP ref map is aligned with the output Q/U map. To make life easier,
this change removes the restriction that the IP ref map must be aligned
with the output Q/U map. Instead, the supplied IP ref map is aligned with
the output map using its WCS Info, and then resampled onto the output
grid.