profile.in: prevent irritating List::Util bug with relocated builds
Bug and fix were originally found when Helen Kirk was having issues
running ORAC-DR.  It occurs when using starlink perl programms in a
relocated starlink build.  I believe it only started occuring after
List::Util functions had been moved into Perl core. We seem to end up
with two versions of List::Util, so this change should cause the
correct one to be at the front of the PERL5LIB variable.

The output seen when this occurs is:

"any" is not exported by the List::Util module
"none" is not exported by the List::Util module