smurf: Allow Q/U maps to be made in focal plane coords
Normally, supplying makemap with Q/U values that are with respect to
focal plane Y axis causes an error to be reported (because of sky
rotation). However, this error can now be suppressed by setting config
parameter "pol2fp" to a non-zero value. This is required for instance
when creating maps of planet data as part of the process of creating
an IP model.