smurf: Do not use static variables in code that is run in a thread
smf_rebin_totmap is called within threaded code from several places. The
"have_azel" static variable may have been read az zero (it's initial
value) by the first threads to start up, before potentially being set to
1 by later threads.

In practice, this will not matter since all it does is to cause the
mapping for the first few time slices to be calculated in a more
long-winded manner than necessary. It affects the speed, but not the
nature of the mapping.