sourceme-example: Add example sourceme file
Added an example "sourceme" file to prepare the build environment
(for csh/tcsh).  Also mentioned the new file in the README.

Note: this file uses a definition of $STARLINK_DIR as the starting
point for defining where everything goes.  It should technically
be $STARCONF_DEFAULT_PREFIX but the Perl module build system requires
$STARLINK_DIR to be defined, so this sourceme file defines it from the
start.