smurf: makemap: allow flagmap parameter to dump individual quality bits
Setting "flagmap=cube" In the config file causes a 3D NDF to be created
in which each plane describes a single quality bit. A plane is created
for every defined quality bit (there are currently 17 of them), plus an
extra plane that counts the number of samples that have no quality bits
set (equivalent to a hits map).