smurf: Merge low priority quality flags so that we have no more than 16
Time series quality values are stored in arrays of type smf_qual_t which
currently is a unsigned short int. This means if we ever go above 16
flags, we will need to double the amount of memory used to hold quality
arrays when makemap is run. This would be very wasteful, considering that
several of the existing quality flags are never used in normal usage of
makemap. So instead, change the low priority flags so that they use a
single generic quality bit, thus allowing us to continue using a short
int for smf_qual_t. They thus become indistinguishable but this does not
matter since they are never used. The two flags merged in this way are:

SMF__Q_FILT - only used if the "new" filtering algorithm is used. This
new algorithm compared unfavourably with the old algorithm, and so is not
used in any of our standard configs.

SMF__Q_BADEF - only used if a sample has a bad optiocal efficiency. But
correcting for optical efficiency was shelved ages ago, and is also not
used in any of our standard configs.