starutil: Search libbsd for strlcpy and strlcat
In PAL, commit 7d6eaa5e70f7edde190f2c41c226c11da02b4a60 added
searching libbsd for strlcpy.  This caused the presence of strlcpy
to be recorded in config.cache, starutil to be built using it
and then sst to fail to build because it tried to use the resultant
starutil library without -lbsd.  (The problem only occurs when
libbsd is installed.)

Therefore update starutil also to search libbsd for consistency
with PAL and add @LIBS@ to starutil_link in case libbsd is found.