one: Remove conf. check for strlcat and strlcpy
ONE has used star_strlcat and star_strlcpy since commit
0978a02061d7048a7182829fd7b041fa6fe5d2ae so it no longer needs
to check for the presence of the system strlcat and strlcpy
functions.  Since it checked for strlcat and strlcpy wihtout
searching libbsd, it would record an absence of strlcat in
config.cache, causing starutil to be built without HAVE_STRLCAT
when it would otherwise have been found in libbsd.