smurf: pol2ip: take account of pointing errors
Previously, the Q, U and I values from which the IP model is derived were
the mean values within cicrular apertures centred on pixel (0,0) - which
is where the source is supposed to be centred. But due to pointing errors,
the source may in fact be upto half a beam width away from (0,0). Now, the
source centroid is found (in the polarised intensity image for Q and U and
in the total intensity image for I), and the aperure is centred on that
position instead.

This should not make much difference for large apertures but will become
increasingly signifificant as the aperture size is reduced.