smurf: pol2ip: Allow model to be based on peak source values
If parameter DIAM is set negative or zero, the IP model is based on the
peak source values rather than the mean values in the source. The peak
values are found using kappa:beamfit on the total intensity and polarised
intensity maps. The peak Q and U values are derived from the peak PI
value on the assumption that the instrumental polarisation is always
parallel to elevation.