kappa: wcsadd can now add a copy of a Frame from another NDF
The two NDFs are aligned properly (e.g. in sky coordinates) so that the
relationship of the new copied Frame to (e.g. ) sky coordinates is the
same in the modified NDF as it was in the reference NDF.

For instance, polpack requires Q and U maps to include POLANAL Frames.
But if you align such a map with an image that has no POLANAL Frame, then
the aligned Q/U map will not have a POLANAL Frame and so cannot be used
by POLPACK. This situation can now be remedied by using wcsadd to add a
POLANAL Frame into the aligned Q/U map, obtaining the POLANAL Frame, and
its relationship to SKY coords, from the original unaligned Q/U map.

% wcsalign qmap ref=imap out=qmap_A
% wcsadd qmap_A maptype=refndf refndf=qmap domain=polanal frame=grid

In the above example, wcsalign copies the WCS FrameSet from imap to
qmap_A, and so the POLANAL Frame present in qmap is lost. The wcsadd
command then adds a POLANAL Frame into qmap_A by copying the Frame from
qmap, aligning qmap and qmap_A in sky coordinates. This means that the
mapping from POLANAL to SKY in qmap will be the same as in qmap_A.