smurf: makemap - make extreme excursions check optional
The new config parameter VALIDATE_SCANS can be set to zero to suppress
the check for crazy scans. This may be necessary if the scan
intentionally spends much time outside the expected map area.

The smf_fix_metadata_scuba2 function is very low level and so it is very
difficult to give it direct access to the config keymap. Instead, a new
facility has been introduced to allow all smurf function to share access
to a set of global values. These global values are held in a KeyMap that
is created within smurf_mon.c. Currently, thus keymap is only available
to the main thread (an error will be reported if it is accessed from a
wortker thread).