Revert "smurf: Add SC2_MAG_X,Y,Z to SCUBA-2 JCMTSTATE structure"
This reverts commit 4ac3cfe43e6bb51ad083df7da1d886c3a24f7f6c.

This change was added by TimJ due to a request from Dan Bintley
to store the high rate magnetometer data into the JCMTSTATE structure.
However, the installed version of this was never updated so the
change was never implemented. Dan Bintley has now stated he doesn't
need this change, so this change is now being reverted