smurf: Add despiking option to WVM smoothing
This adds a simple despiking operation to smf_calc_smoothedwvm
(see fault 20120724.004).  I had hoped to be able to add despiking
to the existing smoothing step, or use the existing despiking routine,
but neither seemed practical.

This needs testing to determine a suitable tolerance, or even if
a fractional tolerance is appropriate.  If we decide to use an absolute
tolerance then it might need to be wavelength-specific as the
opacity values at this point aren't just 225GHz opacities.