smurf: Extract WVM debugging function
Extracted the code to get the raw WVM output for debugging
to form a new function so that the debugging information
can be printed at different stages in the process.  The debugging
information is now sent to stderr to separate it from other
messages (which appear on stdout).  The calls to the new function
are commented out (as was the original debugging code).