smurf: add new python scripts to docs
Without this the commands are not shown in SUN/258 or when using
smurfhelp.

However, something seems to be going wrong with pol2sim.py; prolat
runs happily on pol2sim.py on its own, but when all the python scripts
are catted together, the prolat output doesn't include the docstring
for pol2sim.  This commit will still make the documentation accessible
for pol2scan, pol2ip and pol2stack.