kappa: new command align2d to align images using image features
The align2d command aligns an input 2-d NDF with a reference 2-d NDF
using an affine transformation. The coefficients of the transformation
are determing by minimising the sum of the squared differences between
the reference map and the transformed input map.

It could be better. The initial guess used by the minimisation function
is just a unit transformation. If this is far from the truth, then the
minimisation routine (PDA_LMDIF1) has a lot of trouble trying to find the
minimum. It would be better to provide a better initial guess based on a
phase correlation method such as that described in

http://www-cs.engr.ccny.cuny.edu/~wolberg/pub/icip00.pdf

But sadly, more pressing issues mean that I need to leave this for a
later date.