CCDPACK: Enable the document name to appear in the XML.
Sarah has a script for extracting the information from the package XML
to make a summary web page
(http://www.starlink.ac.uk/docs/starlinksummary.html)
listing packages, libraries, and documents, containing a link to the
documentation and a description.  There are several missing
documentation links and this is the first commit to correct those.