SC/20: Added a..raw file name to the first box and trimmed bluespace at foot.
Also slight alignment of a file name.

Attempts to trim of the extra space at the foot failed using LibreOffice
itself.  This extra space was added for the PNG output in LibreOffice Draw,
which cuts this part off.  Trimming 0.38 inch in the page vertical
dimension to place the arrow head at the foot of the blue region makes
this part narrower than the group part of the graphic, and hence the
downward vector does not align in the figure.  The approach used pdfcrop
to trim 27 points from the foot like below.

  % pdfcrop --margins '0 0 0 -27' sc20_workflow_initial.pdf temp.pdf
  % mv temp.pdf sc20_workflow_initial.pdf