SC/20: Put in some velocity constraints for narrowline and gradient recipes.
The original 20 channels I think was assuming the 0.42 km/s bins
rather than the narrow-band 0.055 km/s, although it may have been
twice the smoothing width in channels.  Surely the important point is
not to use gradient for narrow lines if the smoothing will bury the
line in the background.  For now, until there is a better suggestion,
change from channels to velocities in the guidelines.