ast: Enforce Mapping immutability
In general, Mapping attributes cannot be changed after the Mapping has
been created (the exception to this is the Invert attribute, which can be
changed at any time). However, several of the oldest Mapping classes were
introduced into the AST library before this restriction was enforced. To
reduce the chances of breaking existing software, the attributes of such
Mappings may still be changed, but only for Mapping instances that have
exactly one active reference. In other words, an error will be reported
now if an attempt is made to set or clear an attribute of a Mapping (other
than the Invert attribute) if that Mapping has been cloned. Mappings are
cloned when they are incorporated into another object such as a CmpMap or
FrameSet, or when the astClone function is used.