ast: Avoid modifying SphMap attributes when simplifying a SphMap
The old code changed the attributes of the supplied Mapping as part of
the simplification process. This is dangerous as it could have unforeseen
effects through other pointers to the same mapping, which may be used in
other CmpMaps, FrameSets, etc.

The new code takes a deep copy of the SphMap, before changing its
attributes.