smurf: Only write phs_FPM files in debug mode
This is so that, when you run the FTS-2 pipeline, you don't end up
with a lot of unexpected files in the output directory.  Since the
pipeline doesn't request (or know about) these files, it doesn't
clean them up along with the other intermediate files.  I'm not even
sure what the point of these files is anyway since they seem to be
pretty much the same as what's in the MORE.FTS2.FPM extension of
the main output files.