smurf: Don't read beyond bounds of array
The code for "subsequent inverted times" (i > 0) looks like it
should have been in an "else" block so that it isn't run in the
i == 0 case.  Otherwise it will be accessing array elements
beyond the bounds of the array.  The comments in the code
seem to indicate that this was what was meant to be happening
but there was no "else" block in place to make it actually happen.