smurf: pol2scan: FIx bug in IP beam shape correction
The orientation of the elevation axis within the IP map was being
determined using the Epoch of the IP map itself. But the IP map will have
a different epoch to the POL2 observation, and so does not give the right
position angle for the elevation axis within the output Q/U maps. So
force the IP map epoch to that of the POL2 observation before getting the
position angle of the elevation axis.