kappa: align2d - allow data values to be included in fit
If parameter FITVALS is set TRUE, the relative offset and scale between
the data values in the two images are adjusted, along with the parameters
of the spatial transformation, as part of the fitting process. Any output
NDF generated has this normalisation applied to its pixel values,
resulting in it having the same normalisation as the reference image.