smurf: Use OUT parameter for FTS2SPLIT file names
Now use the entries of the OUT parameter as the base for
output file names rather than ignoring OUT and constructing
the names based on the input file names.  Also simply append
a number to the file (as is done by makecube, for example)
rather than manipulating the subscan number in a confusing
manner.

Note: the FTS-2 pipeline will need to be updated to take
account of the new output files with numbered suffices.