smurf: pol2scan - do not use AST mask if it is very small
For faint sources, the AST mask has very few pixels and so using it to
mask the data prior to using kappa:align2d causes align2d to fail. So use
the whole map if the AST mask is very small (use of the align2d corrlimit
parameter still ensures that noisy background areas are excluded form the
fittign process).