smurf: pol2scan - create maps from separate observations
Separate Q and U maps are made for each observation present in the input
raw data. These are coadded to create the total Q and U maps. Stats for
the maps created from each observation are calculated and displayed, and
an be stored in a an output text file (see parameter OBSTABLE). The
indivisual Q and U maps can be saved for future use (see parameter
MAPDIR).