perlmods: force install LWP::Protocol::https
This module regularly fails one of its tests, particularly on the OSX
nightly build, when attempting to contact apache.org.  As far as we
know, there's never been a real problem with the module in our builds.

It is now explicitly installed, rather than just being pulled in
as part of the Astro::Catalog chain of depenedencies, and it will
be force isntalled without running its tests.