smurf: step fixing boxes should be shorter for low sample rates (e.g. POl2)
Several of the constants used in this routine assume a fixed sample rate
(200 Hz). In pratice, sample rates vary because of down-sampling etc. In
particular, the sample rate for time-streams holding POL-2 Stokes
parameters (produced by calcqu) is 2 Hz, very much lower than normal
SCUBA-2 maps. The usual constant values are inappropriate for this very
low sample rate and cause lots of problems in the step-fixing algorithm.
So we need to change the constants to take the actual sample rate into
account. But such a change could cause potentially important changes to
appear in normal SCUBA_2 maps, which would require a long investigation
to evaluate. We should maybe do this some day, but at the moment the top
priority is getting POL-2 working. So today's code change only changes
the constants for data at sample rates below 20 Hz (corresponding to a
scan speed of 80 arcsec/sec for 4 arcsec pixels). This should fix the
problems for POL-2 but leave normal SCUBA-2 maps unchanged. The values
affected are dcfill, dcmaxwidth, dcsmooth2, dcnlow, dcpeakoff,
dcpeakwidth, dcpeakminwidth. In addition the dcmaxsteps value is also
scaled, but this has to be applied at all sample rates since it is a
configuration parameter and not defined locally. The scaling of
dcmaxsteps leaves the value unchanged for a sample rate of 80 Hz (a scan
speed of 320 arcsec/sec at 4 arcsec pixels).