smurf: Modify pca cleaning to be more robust against no. of threads
Determine the PCA components from the left hand singular vectors
(U) found by smf_svd rather than the right hand (V). U and V should
be the same (because the co-variance matrix is symetric), but U
seems to be less prone to changes as you change the number of
threads.