kappa: look - correct X and Y field width in WLIST mode
X and Y field widths in WLIST and WGLIST format were totally wrong. They
assumes the current Frame was a SkyFrame, which it may not be. It was
really wrong if the current Frame was in fact the PIXEL Frame - in that
it gave a max field width of 3. They now just use twice the value of the
axis Digits attribute.