cupid: Update GaussClumps rejection message
The message printed when using METHOD=GaussClumps and a clump is
rejected previously reported that this was because an edge of the
data array was touched.  However the "thresh" parameter was later
implemented, adding another reason for rejection.  This commit
updates the message to indicate that rejection could have happened
for either reason.  (It would be better to count the number
rejected for each reason and report them separately, but at least
the update combined message is no longer misleading.)