smurf: makemap - allow xxx.zero_niter to be a normalised map change value
For instance, this means that the PCA model can be masked down to a map
change value of 0.2 (say), and then not masked for later iterations. This
is reasonable as by then the vast majority of the astronomical flux
should have been moved out of the time series and into the map. So
leaving the small remaining amount of astronomical signal un-masked should
not affect the PCA model significantly.