smurf: makemap - change default for pca.comfill
By default, do not use the common-mode to fill gaps in the PCA model. The
effects of this parameter need to be more carefully analysed at some
point. Common-mode fill sounds a good idea in theory, but in practice, it
may possibly increase the noise in the final map. Check this out at some
point.