smurf: makemap - allow PCA model to use absolute number of components
If pca.pcathresh is negative, it is now interpreted as a fixed absolute
number of components to include in the PCA model. The same value is used
on all iterations and for all subarrays. This helps convergence. The
maximum number of PCA components is equal to the number of good
bolometers, which is usually around 700. The smaller the value of
pca.pcathresh, the less effect the PCA model has on the residuals. This
causes rapid convergence, but a noisier final map. Setting pca.pcathresh
to a higher value causes the PCA model to remove more from the residuals
on each iteration. This means that convergence takes longer, but the
final map is less noisy. Values between 100 and 200 seem reasonable for
POL2 data.