smurf: makemap - allow user to choose how to mask the normalised map change
Previously, the normalised map change shown at the end of each iteration
was the mean of the normalised pixel changes, averaged over the area of
the AST mask. But if no AST masking is used, then it was averaged over
the whole map. The new config parameter "maptol_mask" allows the user to
select which mask is to be used. The default behaviour is unchanged. But
a value such as "PCA" can be assigned to maptol_mask to force the
normalised map change to be calculated using the PCA mask.