smurf: skyloop - retain SPIKE flags between iterations
Revert the 4-MAR-2014 change, so that the quality array in the cleaned
time series data is again updated by each call to makemap. The 4-MAR-2014
change was made because the time-series quality data was mangled by being
reduced to 8 bits in smf_qual_unmap, and consequently was useless for
later iterations. But this means that spike flags are lost. Now, the new
exportqbits config parameter is used to tell smf_qual_unmap to export the
SPIKE flag. We also export the flags set by the initial cleaning process
(BADDA,BADBOL,DCJUMP,STAT,NOISE,PAD), so that they are available to later
iterations. Since this is only 7 bit, they can be exported without loss
and so can be re-used on subsequent iterations. Things like COM and RING
are recalculated from scratch on each iteration and so do not need to be
exported.