smurf: skyloop - do not de-spike when loading initial image into makemap
When subtracting the AST model specified by an initial sky map (as is
done when makemap is used from skyloop), do not do map-based de-spiking.
This is because the time-series data at that point is the original
cleaned data that includes the noise models (COPM,FLT,etc), and so will
be far removed from the data values implied by the map, making map-based
de-spiking totally unreliable.