smurf: makemap - allow zero_snrlo parameters to be zero
Extending a mask down to an snr of zero is a valid thing to do, but
previously setting zero_snrlo to zero was used to indicate that
zero_snrlo should not be used. This purpose is now achieved instead by
setting zero_snrlo to <undef>, and a value of zero is treated literally.