smurf: makemap - allow convergence to be determined better if no mask is being used
Usually, convegence is indicated by the normalised map change between
iteration falling below the MAPTOL value. The normalised map change is
measured within the source region defined by one of the masks (AST, PCA,
COM, etc) as determined by parameter MAPTOL_MASK. In such cases, if no
masks are being used, then the map change is determined over the whole
map. If the mean map change is used (see paramater MAPTOL_MEAN) then the
mean is usually dominated by the background pixels and is not a good way
of checking for convergence in astronimcal sources. Instead it is better
to use the maximum change rather than the mean change. But the maximum
change is a very noisy statistic. To make it more stable, two new
parameters allow the map change array to be smoothed before being used,
thus making the maximum value representative not just of one pixel but of
a collection of neighbouring pixels, and thus making it less noisy. The
new parameter MAPTOL_BOX can be set to indicate the width so such
smoothing. In addition, the new parameter MAPTOL_HITS can be set to
exclude the intrinsically noisy edges from the smoothing, by specifying a
limiting hits value for data to be included in the map change map.