smurf: A new IP model for use with I maps made from POL2 data
This new IP model is called PL3. And is intended for use with IP refrence
maps made from POL2 data. The PL2 model should be used with IP reference
maps made from non-POL2 data. The default is now for makemap to check the
IP reference map and use PL2 or PL3 as appropriate.