smurf: pol2ip.py: Modify to use I maps derived from the POL2 data
The IREF parameter can now be set to "!", in which case the I value is
derived from the supplied POL2 data.

In addition, the config used is similar to that used by the pol2map
script. Except that no COM model is used as this seems to increase the
scatter in the fit for some unknown reason.