smurf: Another set of changes to the new pol2map script
Main one being that I've reverted back to the co-variance based PCA
cleaner, instead of the new correlation based version. I don't know why,
but the new one seems to give deeper bowls and higher noise than the old
one. Maybe I was just not removing enough components. But the old one
seems to work fine now, so no need for the new one. It seems I was was
wrong to blame the PCA cleaner for the old problems - other changes seem
to have fixed things.