smurf: makemap - Ignore sudden blips when checking convergence
When estimating how many iterations will be needed for convergence, do
not produce an estimate if it is affected by a sudden blip in the rate of
convergence such as caused by config parameters such as xxx.zero_niter.
In such cases the linear regression used to make the estimate will be
unreliable.