smurf: Return BADD where no CSO/WVM fit available
Now return an opacity value of VAL__BADD (rather than 0.0) when
evaluating a CSO or WVM fit and no fit segment covers the requested
time.  Previously SMURF used the value of 0.0 as if it were a
real opacity value, leading to large steps in extinction-corrected
data at the start/end of fitted regions.