smurf: pol2map - allow iqu variances from spread of obs values
If new parameter MAPVAR is set TRUE, the variances in the coadded I, Q
and U maps are formed from the spread of pixel data values in the
individual observation maps. If MAPVAR is FALSE (the default), the
variances in the coadded maps are formed (as before) by propagating the
pixel variance values from the individual observation maps.