smurf: makemap - allow IP correction to tweaked
The new config parameter, "ipoffset", allows the IP model to be raised or
lower by a specified value. It defaults to zero. For instance, the
current "PL3" IP model seems to have an uncertainty of about +/- 0.3%.
Producing three maps with "ipmodel" set to 0.0, -0.3 and +0.3 allows the
possible range of resulting Q and U to be invesigated, giving a handle on
the uncertainties in Q and U caused by the uncertainty in the IP model.