polpack: polwrtcl - allow sky coordinates system to be specified by user
This is so that gaia can be given (e.g.) galactic coordinates in the
catalogue if the display image uses galactic coordinates.

The output celestial coords are still referred to as "RA" and "Dec" even
if they are in fact some other system (which is what RTD/GAIA etc does).