gaia: Correct vector drawing over galactic images
The new method is to leave the contents of the TCL table created by
polwrtcl unchanged, and instead transform the tabulated vector RA/Dec
positions into the coord system of the displayed image each time they are
drawn. There is also the complication that the reference direction used
by the ANG, Q and U columns in the table is galactic north rather than
Dec. So modify polwrtcl so that the created TCL table includes the
position angle of the ref. direction (i.e. angle from Dec to the ref.
direction, in sense of rotation from Dec to RA). Then modify GaiaPolDisp
to rotate the drawn vectors by this amount (assumes constant over the
whole map).

Ugh - that was far more complicated than I expected.