kappa: sqorst: Allow axis scales to be retained
This change allows one or more asterisks to be included in the list
of new pixel scales given by parameter PIXSCALE. An asterisk causes
the original pixel scale to be retained on the associated axis.

For instance, this is useful if you want to change the spatial pixel
scale of a SCUBA-2 map that has a degenerate 3rd pixel axis for
wavelength.