sc22: ensure flowchart is right way up in html output
I manually rotated the figure by 90 degrees (with imagemagick convert)
so that its the right way up in the html, and I added a \latexhtml
command so that in the latex it is rotated by 90 so it can be
larger. A sideways figure in html output seemed non-optimal.

(I couldn't get the PDF to rotate very successfully so I have
converted to .png. As its a flow chart, .pdf format would be fine but
if anyone else changes it then please ensure the image is upright by
default. The angle option to includegraphics isn't obeyed in the html
output.)