ast: Update ast to fix an issue building smurf
The previous version of AST included __FUNCTION__ within the macros used
to invoke AST functions. But __FUNCTION__ is a gcc extension, and so
smurf failed to build (because it uses the -Wpedantic gcc option). This
change to AST replaces __FUNCTION with __func__ (unless __func__is not
available) since __func__ is part of the C99 standard.