Defer error messages in F77 tester so that expected hds errors are hidden
The HDS wrapper attempts to open the supplied file using v4 first. This
will fail if it is a v5 file, with errors being reported. If this
happens, the wrapper annuls the errors and tries to open the file using
v5. But the error messages still get seen by the user unless error deferal
has been switch on.