hds: Update stub for hds-v5 function datLocked so that cary can be built
The C version of ARY does not depend in any direct way on HDS V5.
Therefore it should in principle be possible to build and test C ARY with
HDS-V4. But for this, the stubs provided by v4 for the new V5 locking
functions (which are used by C ARY) need to be kept up to date with
respect to HDS-v5.

The "cary" branch - used for developing C ARY - is based on master,
therefore master needs up-to-date stubs for the hds v5 locking functions.
I presume I don't want to start messing around with HDS on the "cary"
branch?