hds: Stub for datyLocked now assumes the object is locked by the current thread
This is so that testing of C ARY can continue. Since HDS v4 has no object
locking mechanism, there is no way for another thread to lock an object.
SO return 1 to indicate the current thread owns the object. Of course,
hds-v4 should never be used in multi-threaded apps anyway.