pca: Allow _CHAR output parameter values to be greater than 132 characters
The new function SUBPAR_ASS0C is a replacement for SUBPAR_ASSOC that is
intended to be used when storing scalar character strings in output
parameters (i.e. it is called from subpar_put0c). It ensures that the
component of the HDS parameter file in which the string will be stored is
sufficiently long to hold the whole string (excluding trailing spaces)
without truncation. If required, the existing _CHAR component in the parameter file is erased and a new, longer one, created.

This is a bit of a kludge. The whole of SUBPAR seems to be littered with
places where string lengths are hard-wired at 132. It would be better to
parameterise that value, and increase it to some more realistic value
(there already exists the SUBPAR_STRLEN constant, but it seems hardly
ever to be used). I tried doing this but there was clearly subtle issues
going on which I could not spot, since I kept getting lots of stack
smashing and random behaviour. The change implemented in this file seems
a more confined/encapsulated fix that is less likely to have unforeseen
consequences.