smurf: pol2map: Allow catalogue bin size to be specified
New parameter BINSIZE allow the bin size in the vector catalogue to be
specified independently of the pixel size (parameter PIXSIZE). If BINSIZE
is greater than PIXSIZE, the individual observation maps are binned up to
the requested bin size before being coadded and used to form the vector
catalogue. This gets round issues to do with the pixel-to-pixel noise
being correlated due to map-maker artefacts.