hds: Change HDS version parameter to just VERSION
The environment variable was also changed from HDS_VERSION5 to just
HDS_VERSION.

Both versions of the parameter now expect a value of 5 (for HDS
version 5) or 4 (for HDS version 4).  This is so that the parameter
will still be usable for switching back and forth with the same name
whatever the default version happens to be.  (It would also be able
to handle a selection between more than two versions should the need
ever arise in the future.)