hds: Update tuning parameters in SUN/92
The tuning parameter list has been divided into sections based on
the version of the HDS library to which they apply.  (The alternative
would be to leave the list arranged alphabetically but indicate
for each parameter where it applies.)

Some the 64BIT paramters has been added to the documentation, along with
a list of options for the VERSION parameter.