smurf: Add IP models for 450 um data
Configuration parameters relating to the IP model parameters have been
removed, and instead the smf_subip function now chooses the best values
to use based on the FITS headers of the supplied data. Model parameters
suitable for both 450 and 850 data have been included, both with and
without the wind blind, so all these forms of data can now be IP
corrected. The "850 with wind blind" parameters are a little different to
the previous ones, but the corresponding change produced in IP is
insignificant compared to the noise.

See article "IP model without the wind blind" in the "Data Reduction And
Analysis" section of the POL2 commissioning wiki.