cary: Allow forcing of old ORIGIN data type
When we switch to 8 byte integer pixel indexing, the ORIGIN component of
new NDFs needs to change to 8 bytes too in order to accomodate
potentially huge origin values. But we need to make it possible for such
future systems to export NDFs in a form that can be read by older systems
that expect 4 byte origins. This change allows user of a new system to
force 4 byte ORIGINs by defining the environment variable NDF_SHORTORIGIN.