cary: Allow old ARY libraries to read new NDFs
When checking the data type of the ORIGIN component of an NDF, allow the
type to be either _INTEGER or _INT64 explicitly (i.e. use "_INT64" in the
check instead of "HDS_DIM_TYPE"). This is because people with older
starlink installations may still have HDS_DIM_TYPE set to _INTEGER, and
we want such people to be able to read new NDFs that use _INT64 for
ORIGIN. This relies on the automatic type conversion provided by HDS to
convert _INT64 to _INTEGER when the ORIGIN values are actually accessed.