hds: Update sun92 for v5/6
SUN/92 was way out of date. It still is not fully up to date. All I've
done is include a few sections about the change to HDF5. Things like the
list of error codes (who uses that??) are still not right, but it's
better than it was and does at least now indicate that there has been a
major change. More corrections can be made if HDS suddenly becomes of
wider interest. Otherwise, we can carry on as we have done for years.

The main pain is the fact that the version numbers are different for the
library and the data format. So the data format is moving to V5 but the
library is moving to V6. Maybe we should have just skipped data format V5
and gone straight from V4 to V6 so that the data format and library are
in step with each other.