kaplibs: kappa:align2d: Use SNRs to weight residuals
The KPG1_ALIGN function (used primarily by kappa:align2d) used the
reiprocal of the variances to weight the residuals when doing the least
squares fit. But the variance around bright point sources is often very
high causing them to be largely ignored, when in fact point sources are
good for determing alignment. So use SNR instead of reciprocal variance
for the weights.